+++ tests/testthat/test-std_center_scale_args.R
@@ -1,5 +1,4 @@
 if (require("testthat")) {
-
   d <- data.frame(a = 1:5, b = 21:25, c = 41:45)
 
   test_that("standardize", {