+++ NEWS.md
@@ -3,6 +3,17 @@
 - `reshape_ci()` gains a `ci_type` argument, to reshape data frames where 
   CI-columns have prefixes other than `"CI"`.
 
+- `center()` gains the arguments `force` and `reference`, similar to
+  `standardize()`.
+
+- The functionality of the `append` argument in `center()` and `standardize()` 
+  was revised. This made the `suffix` argument redundant, and thus it was 
+  removed.
+
+- Fixed issue in `standardize()`.
+
+- Fixed issue in `data_findcols()`.
+
 # datawizard 0.2.1
 
 - Exports `plot` method for `visualisation_recipe()` objects from `{see}`